+++ docs/CHANGE_LOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0.6.0:
 
+#### **Binding Options / Function Options:**
+- Added a new binding/option called "allowCanceling", which states the watch can be canceled (defaults to true).
+- Added a new binding/option called "allowPausing", which states the watch can be paused (defaults to true).
+
 #### **Public Functions:**
 - Added new public function "pauseWatches()", which is used to pause all the watches for a specific number of milliseconds.
 - Added new public function "resumeWatches()", which is used to resume all the watches currently paused.